refactor: Change color usage into using defined color variable
