+++ src/components/index.ts
@@ -3,11 +3,5 @@ export { Textarea } from './ui/textarea';
 export { Button } from './ui/button';
 export { FileInput } from './ui/file-input';
 export { Input } from './ui/input';
-
-
-
-
-
-
-// Nando
 export { Stepper } from "./ui/stepper"
+export { CheckIcon } from './ui/check-icon';
