+++ src/components/ui/check-icon.tsx
@@ -1,0 +1,14 @@
+export const CheckIcon: React.FC<{ className?: string }> = () => {
+  return (
+    <svg
+      className= "w-6 h-6"
+      fill="none"
+      stroke="currentColor"
+      strokeWidth={8}
+      viewBox="0 0 24 24"
+      xmlns="http://www.w3.org/2000/svg"
+    >
+      <path strokeLinecap="round" strokeLinejoin="round" strokeWidth="2" d="M5 13l4 4L19 7" />
+    </svg>
+  );
+};
